requirements.txt
streamlit
mysql-connector-python
openai